social_media_api/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.15.0,<4.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
gunicorn>=21.0.0
whitenoise>=6.5.0